requirements.txt
selenium==4.9.0
python-telegram-bot==13.15